python/requirements.txt
spleeter
openai-whisper
ffmpeg-python
numpy
torch
torchaudio
